feat: add VoteButton and VoteDialog components; refactor imports in MissionCard and MissionCreateButton
